refactor: reorganize project structure and clean up root directory
Major changes:
1. Directory Structure Optimization
   - Create UI/ directory to consolidate all PySide6 logic, views, and resource definitions.
   - Create assets/img/ directory to unify storage of icons, previews, and third-party logos.
2. Logic Decoupling and Refactoring
   - Add Util/Path.py for unified absolute path conversion logic.
   - Refactor UI modules into a package structure using relative imports.
   - Remove redundant code from KeymouseGo.py and Plugin/Manager.py.
3. Compatibility Updates
   - Update image reference paths in README.md and README_en-US.md.
   - Update UI/assets.qrc paths and re-generate assets_rc.py.
   - Optimize .gitignore rules.
