+++ UI/Internationalizing.pro
@@ -1,3 +1,3 @@
 SOURCES         = UIView.py UIFileDialogView.py
 FORMS           = UIView.ui UIFileDialogView.ui
-TRANSLATIONS    = assets/i18n/en.ts assets/i18n/zh-cn.ts
+TRANSLATIONS    = ../assets/i18n/en.ts ../assets/i18n/zh-cn.ts ../assets/i18n/zh-tw.ts
