+++ UI/assets.qrc
@@ -1,5 +1,5 @@
 <RCC>
   <qresource prefix="pic">
-    <file>Mondrian.png</file>
+    <file>../assets/img/Mondrian.png</file>
   </qresource>
 </RCC>
