+++ UI/assets_rc.py
@@ -1,6 +1,6 @@
 # Resource object code (Python 3)
 # Created by: object code
-# Created by: The Resource Compiler for Qt version 5.15.2
+# Created by: The Resource Compiler for Qt version 6.5.3
 # WARNING! All changes made in this file will be lost!
 
 from PySide6 import QtCore
@@ -27,6 +27,14 @@ qt_resource_name = b"\
 \x00\x00v\xf3\
 \x00p\
 \x00i\x00c\
+\x00\x06\
+\x06\x8a\x9c\xb3\
+\x00a\
+\x00s\x00s\x00e\x00t\x00s\
+\x00\x03\
+\x00\x00p7\
+\x00i\
+\x00m\x00g\
 \x00\x0c\
 \x06\x0d\xc0G\
 \x00M\
@@ -38,18 +46,13 @@ qt_resource_struct = b"\
 \x00\x00\x00\x00\x00\x00\x00\x00\
 \x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
 \x00\x00\x00\x00\x00\x00\x00\x00\
-\x00\x00\x00\x0c\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-\x00\x00\x01\x80\x8e\x02\xff\x81\
+\x00\x00\x00\x0c\x00\x02\x00\x00\x00\x01\x00\x00\x00\x03\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x00\x1e\x00\x02\x00\x00\x00\x01\x00\x00\x00\x04\
+\x00\x00\x00\x00\x00\x00\x00\x00\
+\x00\x00\x00*\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
+\x00\x00\x01\x93\xf2\xd1\x8f\xcf\
 "
-
-icon = b"\
-iVBORw0KGgoAAAANSUhEUgAAAC\
-AAAAAgCAYAAABzenr0AAAABHNCSVQICAgIfAhkiAAAAHFJREFUWIXt1j\
-sKgDAQRdF7xY25cpcWC60kioI6Fm/ahHBCMh+BRmGMnAgEWnvPpzK8dv\
-rFCCCAcoD8og4c5Lr6WB3Q3l1TBwLYPuF3YS1gn1HphgEEEABcKERrGy\
-0E3B0HFJg7C1N/f/kTBBBA+Vi+AMkgFEvBPD17AAAAAElFTkSuQmCC\
-"
-
 
 def qInitResources():
     QtCore.qRegisterResourceData(0x03, qt_resource_struct, qt_resource_name, qt_resource_data)
